driver and controller SET, GetSet DONE
Postman example URL to try:

{{base_url}}/api/sets/get?RoutineId=10&ExerciseId=6

+++ backendGo/StructModels/Set.go
@@ -1,10 +1,10 @@
 package structmodels
 
 type Set struct {
-	ID         int `json:"id"`
-	RoutineID  int `json:"routine_id"`
-	ExerciseID int `json:"exercise_id"`
-	Set_number int `json:"set_number"`
-	Reps       int `json:"reps"`
-	Weight     int `json:"weight"`
+	ID         int     `json:"id"`
+	RoutineID  int     `json:"routine_id"`
+	ExerciseID int     `json:"exercise_id"`
+	Set_number int     `json:"set_number"`
+	Reps       int     `json:"reps"`
+	Weight     float32 `json:"weight"`
 }
